requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.1.0
python-multipart==0.0.9
cryptography==42.0.5